Notify the flow thread about anonymous block removal in time.
CL:3432406 didn't get this entirely correct. For an anonymous block that
is going to be coalesced into a sibling anonymous block (because there's
no longer anything separating them), it's too late to notify the flow
thread when the inline content has moved into the other anonymous block,
because there may be column spanners in the inline content (now that we
have block-in-inline support). And the flow thread would be confused and
fail to find the column set to which the anonymous block belongs.

Bug: 1292004
Change-Id: I60a05891bdede314023941cd5ef4ca53b8c8fb91